+++ scripts/deploy.sh
@@ -1,0 +1,25 @@
+#!/usr/bin/env sh
+
+# Make sure the script throws the error encountered
+set -e
+
+# Generate static files
+pnpm docs:build
+
+# Go to the generated folder
+cd docs/.vitepress/dist
+
+# If you are publishing to a custom domain
+# echo 'github.saqqdy.com' > CNAME
+
+git init
+git add .
+git commit -m 'deploy'
+
+# if publishing to https://<USERNAME>.github.io
+# git push -f git@github.com:<USERNAME>/<USERNAME>.github.io.git main
+
+# Replace <USERNAME> with your own Github username and <REPO> with the repository name
+git push -f https://github.com/saqqdy/rollup-plugin-replace-shebang.git main:gh-pages
+
+cd -